Rename function
Instead of moving the cookie-setting process into the login function, I
decided to keep them in the route and renamed the origin function as a
simple bool function. This is because the cookie-setting depends on
existing response, but making a response in the util function "login"
seems to be too much: the one that makes a response should be the route
itself.